requirements.txt
python-cinderclient>=1.6.0,!=1.7.0,!=1.7.1  # Apache-2.0
python-glanceclient>=2.3.0,!=2.4.0  # Apache-2.0
python-keystoneclient>=2.0.0,!=2.1.0  # Apache-2.0
python-neutronclient>=5.1.0  # Apache-2.0
python-novaclient>=2.29.0,!=2.33.0  # Apache-2.0
leveldb>=0.194
pbr>=1.6
setuptools>=16.0
